+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Unreleased Changes
 * `MultiSelectDropDown component <https://github.com/anvilistas/anvil-extras/pull/44>`_
 * `@wait_for_writeback decorator <https://github.com/anvilistas/anvil-extras/pull/50>`_
 * `Quill component <https://github.com/anvilistas/anvil-extras/pull/52>`_
+* `Switch component <https://github.com/anvilistas/anvil-extras/pull/31>`_
 
 
 v1.1.0 27-Mar-2021